Adds an option to set the maximum number of frames the program hashes per video and adds an option to set the percent of the beginning and end of each video to skip when hashing.

+++ src/DecodingFrames.h
@@ -7,7 +7,7 @@
 
 std::vector<uint64_t> decode_and_hash(
     std::string const& file,
-    double skip_pct,
+    int skip_pct,
     int duration_s,
     int max_frames,
     std::function<void(int)> const& on_progress = {});
